Create README.md
Update: Token maximum supply

Increased the maximum supply of Voters Trust Token (VTT) to 7.4 billion tokens to accommodate a larger user base and future expansion. This quantity allows for a more inclusive approach to token distribution, ensuring that every voter can participate in the election process with confidence. The launch date remains set for October 1st, 2023, aligning with Nigeria's Independence Day, signifying the significance of democracy and elections in the country.

The decision to adjust the maximum supply was reached after careful consideration of the potential user base and to maintain the token's scarcity and value. Additionally, this change ensures that VTT remains a reliable and sustainable solution for improving transparency and trust in voting systems.

Special thanks to Emmanuel Oche Samuel, Stephanie Oloko, Jeremiah Benjamin Osigbemeh, Francis Ofomata, and Joy Enweliku for their contributions and valuable insights throughout the decision-making process.

Related issue: #123